List bug
The build failed because CMake couldn't find the CMakeLists.txt file. This is because I'm running the command from the base directory /Users/dante/Desktop/reacthreejs/SCT, but CMake needs to be run from the build directory. Since I cannot cd, I need to specify the full path to the CMakeLists.txt file.